Split up the machine vertex (currently broken somewhere)
